0.11

+++ src/main/java/com/smart/house/app/dto/HouseSpec/HouseSpecRequestDto.java
@@ -10,7 +10,7 @@ import java.math.BigDecimal;
 @Getter
 @Setter
 public class HouseSpecRequestDto {
-    private Integer floors; // юзать тип покороче
+    private Byte floors;
 
     private BigDecimal totalCost;
 }
