+++ src/main/java/com/smart/house/app/dto/HouseSpec/HouseSpecResponseDto.java
@@ -12,7 +12,7 @@ import java.math.BigDecimal;
 @Setter
 public class HouseSpecResponseDto {
 
-    private Integer floors;
+    private Byte floors;
 
     private BigDecimal totalCost;
 }
